Move message marshaling code to separate file
